requirements.txt
# Core dependencies
torch
torchvision
torchaudio

# Additional dependencies for image processing and handling
Pillow
opencv-python-headless
scikit-learn

# Web framework for the app
streamlit
python-dotenv

# MongoDB integration
pymongo

# Additional utilities
numpy
matplotlib
requests
pandas

# If you're also using TensorFlow
tensorflow


